Reproduce IDE error
With this set up we have an expect/actual interface where the actual interface in androidMain defines a function. Then we have an interface in commonMain that extends the expect/actual interface and uses generics. Finally, a class implementation in androidMain extends the intermediate interface.

The issue is that from the perspective of the implementation, the IDE does not recognize the interface's function.

One thing to note is if we remove the generics from the intermediate interface, the issue resolves itself.

Also of note, this compiles and runs without issue.